requirements.txt
absl-py
anyio
argon2-cffi
argon2-cffi-bindings
astor
async-generator
atomicwrites
attrs
Babel
backcall
bleach
certifi
cffi
charset-normalizer
colorama
contextvars
cycler
dataclasses
deap
decorator
defusedxml
Django
entrypoints
future
gast
google-pasta
grpcio
h5py
idna
immutables
importlib-metadata
importlib-resources
iniconfig
ipykernel
ipython
ipython-genutils
jedi
Jinja2
joblib
json5
jsonschema
jupyter-client
jupyter-core
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-server
Keras
Keras-Applications
Keras-Preprocessing
kiwisolver
Markdown
MarkupSafe
matplotlib
mistune
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numpy
opencv-python
packaging
pandas
pandocfilters
parso
pickleshare
Pillow
pluggy
prometheus-client
prompt-toolkit
protobuf
py
pycparser
Pygments
PyMySQL
pyparsing
pyrsistent
pyswarms
pytest
python-dateutil
pytz
pywin32
pywinpty
PyYAML
pyzmq
requests
scikit-learn
scipy
seaborn
Send2Trash
six
sklearn-extensions
sklearn-genetic
sniffio
tensorboard
tensorflow
tensorflow-estimator
termcolor
terminado
testpath
tk
tomli
tornado
tqdm
traitlets
typing_extensions
urllib3
wcwidth
webencodings
websocket-client
Werkzeug
wrapt
zipp
